pyproject.toml
[project]
name = "puch-hackathon-autolysis"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "chardet>=5.2.0",
    "fastmcp>=2.11.2",
    "httpx>=0.28.1",
    "matplotlib>=3.10.5",
    "pandas>=2.3.1",
    "python-dotenv>=1.1.1",
    "scikit-learn>=1.7.1",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "tabulate>=0.9.0",
]
